pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "IlolG"
version = "0.1.0"
description = "Bot discord pour suivre les statistiques et le gain de LP sur League of Legends"
authors = [
    { name = "LjeAlexis", email = "alexislajugiepro@gmail.com" }
]
license = { text = "MIT" }
dependencies = [
    "discord.py>=2.0.0",
    "python-dotenv>=0.21.0",
    "requests>=2.28.0",
    "apscheduler>=3.10.0"
]

[project.urls]
Homepage = "https://votre-repo.github.io"
Source = "https://github.com/votre-repo"

[tool.setuptools.packages.find]
where = ["ilolg"]

[project.scripts]
lol-discord-bot = "ilolg.bot:main"
